# [Misc] Set correct OMP_NUM_THREADS for executors