@cyb70289 2025-08-13
> Hi @cyb70289 Perhaps you can try to enable and verify the core binding feature in #21032 for ARM and avoid modify the executor.

Thanks @bigPYJ1151 , I'm not sure if #21032 can fix my issue. I tested on latest master branch with the command line listed above. I want to run on CPU0~15, besides VLLM_CPU_OMP_THREADS_BIND="0-15", anything else shall I configure?