@bigPYJ1151 2025-08-13
@cyb70289 Because the feature is only enabled on x86 and POWER, we didn't verify it on ARM. 
It will be great if you can help to check :)

https://github.com/vllm-project/vllm/blob/e18859298d109870b22cb5b8672d1078818e268d/vllm/v1/worker/cpu_worker.py#L44-L57