@cyb70289 2025-08-13
@bigPYJ1151 , I will test on Arm server, is this feature only works when VLLM_CPU_OMP_THREADS_BIND is not set?