@bigPYJ1151 2025-08-13
@cyb70289 Yes, if set ```VLLM_CPU_OMP_THREADS_BIND``` explicitly, CPU worker will not auto select CPU cores. 
It doesn't matter, for SMT-1 just set ```cpu_selector``` of ```_get_autobind_cpu_ids``` as ```lambda cpus: cpus```.